pyproject.toml
[project]
name = "infinity"
version = "0.1.0"
description = "SDK for creating Hugging Face Inference Endpoints deployments"
readme = "README.md"
authors = [
    { name = "Morgan Funtowicz", email = "morgan@hf.co" }
]
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "packaging>=24.2",
]

[project.optional-dependencies]
openai = [
    "fastapi>=0.115.11",
    "python-multipart>=0.0.20",
]
audio = [
    "librosa>=0.10.2.post1",
]
vllm = [
    "vllm>=0.7.3",
]
cli = [
    "typer>=0.15.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/infinity"]
